Adding FilesystemTarget unit-test
