--- bin/diff_target.py
@@ -1,145 +1,0 @@
-# Copyright: Hallux team, 2023
-
-# We have several options where (and how) to send found fixes.
-# 1. Most primitive one - it simply to re-write local files
-# 2. Little more complicated - rewrite local files, but also make git commit on every change
-# 3. Submit all changes onto Github Web Interface as proposals/suggestions
-#    This last one requires some extra configuring to be done
-# All those approaches are hidden after DiffTarget interface
-
-from __future__ import annotations
-from abc import ABC, abstractmethod
-import subprocess
-import os
-from pathlib import Path
-from github import Github, Repository, PullRequest
-from file_diff import FileDiff
-
-
-# Interface for DiffTarget implementations
-class DiffTarget(ABC):
-    @abstractmethod
-    def apply_diff(self, diff: FileDiff) -> None:
-        pass
-
-    @abstractmethod
-    def revert_diff(self) -> None:
-        pass
-
-    @abstractmethod
-    def commit_diff(self) -> bool:
-        """
-        :return: true if issues are solved independently
-                (i.e. need to increment issue_index in the IssueSolver)
-        """
-        pass
-
-
-# Saves diffs directly into local filesystem
-class FilesystemTarget(DiffTarget):
-    def __init__(self):
-        self.existing_diff: FileDiff | None = None
-
-    def apply_diff(self, diff: FileDiff) -> None:
-        if self.existing_diff is not None:
-            raise Exception("FilesystemTarget: Cannot apply new diff, before last one committed or reverted")
-
-        self.existing_diff = diff  # save new diff in the memory
-
-        with open(diff.filename, "wt") as file:
-            for line in range(0, diff.start_line + 1):
-                file.write(diff.filelines[line] + "\n")
-
-            for code_line in diff.proposed_lines[1:]:
-                file.write(code_line + "\n")
-
-            for line in range(diff.end_line, len(diff.filelines)):
-                file.write(diff.filelines[line] + "\n")
-
-            file.close()
-
-    def revert_diff(self) -> None:
-        if self.existing_diff is not None:
-            with open(self.existing_diff.filename, "wt") as file:
-                for line in self.existing_diff.filelines:
-                    file.write(line + "\n")
-            self.existing_diff = None
-
-    def commit_diff(self) -> bool:
-        self.existing_diff = None
-        return False
-
-
-# Saves fixes into local git commits
-class GitCommitTarget(FilesystemTarget):
-    def __init__(self):
-        FilesystemTarget.__init__(self)
-
-    def apply_diff(self, diff: FileDiff) -> None:
-        FilesystemTarget.apply_diff(self, diff)
-
-    def revert_diff(self) -> None:
-        FilesystemTarget.revert_diff(self)
-
-    def commit_diff(self) -> int:
-        curr_dir: str = os.getcwd()
-        git_dir: str = str(Path(self.existing_diff.filename).parent)
-        os.chdir(git_dir)
-        subprocess.check_output(["git", "add", os.path.relpath(self.existing_diff.filename, start=git_dir)])
-        subprocess.check_output(["git", "commit", "-m", f'"{self.existing_diff.description}"'])
-        FilesystemTarget.commit_diff(self)
-        os.chdir(curr_dir)
-        return 0
-
-
-# Saves Issue Fixes as Github proposals
-class GithubProposalTraget(FilesystemTarget):
-    def __init__(self, config: dict, PullRequestID):
-        FilesystemTarget.__init__(self)
-        self.config = config
-        self.PullRequestID = PullRequestID
-        if "GITHUB_TOKEN" not in os.environ.keys():
-            print("GITHUB_TOKEN is not provided properly")
-            exit(6)
-
-        base_url = self.config["base_url"] if "base_url" in self.config else "https://api.github.com"
-        self.github = Github(os.environ["GITHUB_TOKEN"], base_url=base_url)
-
-        if "repo_name_or_id" not in self.config:
-            print("repo_name_or_id is not provided properly")
-            exit(6)
-        repo_name_or_id = self.config["repo_name_or_id"]
-        self.repo: Repository = self.github.get_repo(repo_name_or_id)
-        self.pull_request: PullRequest = self.repo.get_pull(self.PullRequestID)
-        print(self.pull_request.title)
-
-        if self.pull_request.closed_at is not None:  # self.pull_request.is_merged :
-            print(f"Pull Request {self.PullRequestID} is either closed or merged already")
-            exit(6)
-        latest_github_sha: str = self.pull_request.head.sha
-        git_log = subprocess.check_output(["git", "log", "--pretty=oneline"]).decode("utf8")
-        local_git_sha = git_log.split("\n")[0].split(" ")[0]
-
-        if local_git_sha != latest_github_sha:
-            print(
-                f"Local git commit `{local_git_sha}` and head from pull-request `{latest_github_sha}` do not coincide!"
-            )
-            exit(6)
-
-    def apply_diff(self, diff: FileDiff) -> None:
-        FilesystemTarget.apply_diff(self, diff)
-
-    def revert_diff(self) -> None:
-        FilesystemTarget.revert_diff(self)
-
-    def commit_diff(self) -> bool:
-        commit = self.repo.get_commit(self.pull_request.head.sha)
-        body = self.existing_diff.description + "\n```suggestion\n"
-        for line in self.existing_diff.proposed_lines:
-            body = body + line + "\n"
-        body = body + "\n```"
-        self.pull_request.create_review_comment(
-            body=body, commit=commit, path=self.existing_diff.filename, line=self.existing_diff.issue_line
-        )
-        FilesystemTarget.revert_diff(self)
-        return True
