+++ bin/python/python.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 from query_backend import QueryBackend
-from diff_target import DiffTarget
+from targets.diff_target import DiffTarget
 from code_processor import CodeProcessor, set_directory
 from pathlib import Path
 from python.ruff import Ruff_IssueSolver
