+++ hallux-test.sh
@@ -2,6 +2,10 @@
 
 source ./activate.sh
 
-python -m pytest tests/integration/
-
+python -m pytest -v \
+    --cov-branch \
+    --cov-report=term \
+    --cov=bin \
+    tests/integration \
+    tests/unit
