+++ requirements.txt
@@ -1,4 +1,5 @@
 pytest==7.3.1
+pytest-cov==4.1.0
 ruff==0.0.272
 black==23.3.0
 openai==0.27.8
