+++ tests/test-project1/cpp/test_cpp_project.cpp
@@ -20,5 +20,3 @@ int main(int argc, char* argv[])
 
   return 0
 }
-
-
